Generate Swagger docs as part of the Next.js build process
